Remove uses of Thread.Abort
Update most "runner" classes to use async tasks and cancellation tokens
to gracefully abort instead of depending on Thread.Abort() was is not
supported on pure .NET Core platforms.

Currently Controller.cs is still missing some functionality.  I'm not
familiar with how CoCoL handles creating tasks, so it's not clear to me
how the tasks created by that controller are managed, but right now that
means that there's currently no way to abort the controller tasks.
Suggestions on where to create or pass in a cancellation token would
be appreciated.